muzik/airflow/requirements.txt
apache-airflow-providers-mysql
mysql-connector-python
Faker
